# Improve performance of Semantic.Git.lsTree.