--- a/semantic.cabal
+++ b/semantic.cabal
@@ -305,8 +305,8 @@ library
                      , reducers ^>= 3.12.3
                      , semigroupoids ^>= 5.3.2
                      , split ^>= 0.2.3.3
+                     , streaming-attoparsec ^>= 1.0.0.1
                      , streaming-process ^>= 0.1
-                     , streaming-utils ^>= 0.2
                      , stm-chans ^>= 3.0.0.4
                      , template-haskell ^>= 2.14
                      , time ^>= 1.8.0.2