--- a/src/Semantic/Git.hs
+++ b/src/Semantic/Git.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE DeriveAnyClass #-}
+{-# OPTIONS_GHC -O1 #-}
+
 module Semantic.Git
   ( -- Primary (partial) API for cmd line git
     clone
@@ -15,14 +18,22 @@ module Semantic.Git
   , parseEntry
   ) where
 
-import Control.Monad.IO.Class
-import Data.Attoparsec.Text   (Parser)
-import Data.Attoparsec.Text   as AP
-import Data.Char
-import Data.Either            (fromRight)
-import Data.Text              as Text
-import Shelly                 hiding (FilePath)
-import System.IO              (hSetBinaryMode)
+import Prologue
+
+import           Control.Monad.IO.Class
+import qualified Data.ByteString.Char8 as BC
+import           Data.Attoparsec.ByteString.Char8 (Parser)
+import           Data.Attoparsec.ByteString.Char8 as AP
+import qualified Data.Attoparsec.ByteString.Streaming as AP.Stream
+import qualified Data.ByteString.Streaming.Char8 as ByteStream
+import qualified Streaming.Process
+import           Data.Char
+import           Data.Either (fromLeft, fromRight)
+import           Data.Text as Text
+import qualified Data.Text.Encoding as Text
+import           Shelly hiding (FilePath)
+import qualified Streaming.Prelude as Stream
+import           System.IO (hSetBinaryMode)
 
 -- | git clone --bare
 clone :: Text -> FilePath -> IO ()
@@ -32,24 +43,31 @@ clone url path = sh $ do
 -- | git cat-file -p
 catFile :: FilePath -> OID -> IO Text
 catFile gitDir (OID oid) = sh $ do
-  run "git" ["-C", pack gitDir, "cat-file", "-p", oid]
+  run "git" ["-C", pack gitDir, "cat-file", "-p", Text.decodeUtf8 oid]
 
 -- | git ls-tree -rz
 lsTree :: FilePath -> OID -> IO [TreeEntry]
-lsTree gitDir (OID sha) = sh $ parseEntries <$> run "git" ["-C", pack gitDir, "ls-tree", "-rz", sha]
+lsTree gitDir (OID sha) = Streaming.Process.withStreamingOutput lsproc go
+  where
+    lsproc = Streaming.Process.proc "git" ["-C", gitDir, "cat-file", "-p", BC.unpack sha]
+    go :: ByteStream.ByteString IO () -> IO [TreeEntry]
+    go
+      = fmap (fromLeft [] . fst)
+      . AP.Stream.parse everything
 
 sh :: MonadIO m => Sh a -> m a
 sh = shelly . silently . onCommandHandles (initOutputHandles (`hSetBinaryMode` True))
 
 -- | Parses an list of entries separated by \NUL, and on failure return []
-parseEntries :: Text -> [TreeEntry]
+parseEntries :: BC.ByteString -> [TreeEntry]
 parseEntries = fromRight [] . AP.parseOnly everything
-  where
-    everything = AP.sepBy entryParser "\NUL" <* "\NUL\n" <* AP.endOfInput
+
+everything :: Parser [TreeEntry]
+everything = AP.sepBy entryParser "\NUL" <* "\NUL\n" <* AP.endOfInput
 
 -- | Parse the entire input with entryParser, and on failure return a default
 -- For testing purposes only
-parseEntry :: Text -> Either String TreeEntry
+parseEntry :: BC.ByteString -> Either String TreeEntry
 parseEntry = AP.parseOnly (entryParser <* AP.endOfInput)
 
 -- | Parses a TreeEntry
@@ -58,14 +76,14 @@ entryParser = TreeEntry
   <$> modeParser <* AP.char ' '
   <*> typeParser <* AP.char ' '
   <*> oidParser <* AP.char '\t'
-  <*> (unpack <$> AP.takeWhile (/= '\NUL'))
+  <*> (BC.unpack <$> AP.takeWhile (/= '\NUL'))
     where
       typeParser = AP.choice [BlobObject <$ "blob", TreeObject <$ "tree", OtherObjectType <$ AP.takeWhile isAlphaNum]
       modeParser = AP.choice [NormalMode <$ "100644", ExecutableMode <$ "100755", SymlinkMode <$ "120000", TreeMode <$ "040000", OtherMode <$ AP.takeWhile isAlphaNum]
       oidParser = OID <$> AP.takeWhile isHexDigit
 
-newtype OID = OID Text
-  deriving (Eq, Show, Ord)
+newtype OID = OID BC.ByteString
+  deriving (Eq, Show, Ord, Generic, NFData)
 
 data ObjectMode
   = NormalMode
@@ -73,13 +91,13 @@ data ObjectMode
   | SymlinkMode
   | TreeMode
   | OtherMode
-  deriving (Eq, Show)
+  deriving (Eq, Show, Generic, NFData)
 
 data ObjectType
   = BlobObject
   | TreeObject
   | OtherObjectType
-  deriving (Eq, Show)
+  deriving (Eq, Show, Generic, NFData)
 
 data TreeEntry
   = TreeEntry
@@ -87,4 +105,4 @@ data TreeEntry
   , treeEntryType :: ObjectType
   , treeEntryOid  :: OID
   , treeEntryPath :: FilePath
-  } deriving (Eq, Show)
+  } deriving (Eq, Show, Generic, NFData)