--- a/src/Semantic/Git.hs
+++ b/src/Semantic/Git.hs
@@ -59,7 +59,7 @@ parseEntries :: BC.ByteString -> [TreeEntry]
 parseEntries = fromRight [] . AP.parseOnly everything
 
 everything :: Parser [TreeEntry]
-everything = AP.sepBy entryParser "\NUL"
+everything = AP.sepBy entryParser "\NUL" <* optional "\NUL" <* AP.endOfInput
 
 -- | Parse the entire input with entryParser, and on failure return a default
 -- For testing purposes only