- [x] Depends on #286.

As currently implemented, `lsTree` calls out to `shelly` to invoke `git ls-tree`, which loads the entire response into memory as a strict `Text` before yielding a final result. That response is immediately deconstructed by being fed into an Attoparsec parser. While this works, it’s not as efficient as it could be given large repos (for example, Kubernetes yields 18000 results for this invocation).

A better solution is to use `streaming-process` and `streaming-attoparsec` to stream the output of this process directly into the Attoparsec parser. This also avoids the overhead of converting from the system-level `ByteString` to the `Text` in which `shelly` operates, and (I believe) makes the parser slightly more efficient, as `ByteString`s are easier to traverse than `Text`.

My testing over `kubernetes` yielded a ~25% speedup. Before:
```
benchmarked without-streaming
time                 161.6 ms   (158.9 ms .. 164.3 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 160.9 ms   (159.9 ms .. 162.6 ms)
std dev              2.165 ms   (1.285 ms .. 3.131 ms)

```

After:
```
benchmarked with streaming
time                 123.8 ms   (117.5 ms .. 127.7 ms)
                     0.997 R²   (0.988 R² .. 1.000 R²)
mean                 131.8 ms   (129.0 ms .. 135.8 ms)
std dev              5.846 ms   (4.254 ms .. 7.176 ms)
```

This lies on a hot path for other services, so it’s important that we get this right, but I feel confident that we have enough unit tests to catch any bugs that might occur (see #286).