@patrickt 2019-09-29
Closing this until we get a handle on the UTF-8 situation.